Add TOD sync logic
Fixes #3

Due to drift, adjustments in system clock, or an inaccurate TSC
frequency the local TOD clock may diverge too far from the system TOD
clock. This solution makes a best effort to keep the clocks in sync by
calling into the kernel if it's been 1ms or greater since last sync.
We can still return inaccurate results if the system clock changes
inside that window or if the calculated TSC frequency is inaccurate.

The ideal method would map a read-only page in userland that contains
the kernel's TOD base (see clock()).

I'm still seeing up to 10us delta even with synching. I haven't
determined the exact cause but my hunch is that it is either the TOD
adjustment mechanism found in clock() or an inaccurate TSC frequency.